requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
sqlmodel
python-dotenv
pyjwt
passlib[argon2]
aiosqlite
pydantic-settings
asyncpg
